bot/requirements.txt
python-telegram-bot>=20.0
requests>=2.31.0
pillow>=10.0.0